python/requirements.txt
dnspython==2.2.1
mysql-connector-python==8.0.30
paho-mqtt==1.6.1
protobuf==3.20.1
